+++ src/main/java/de/samply/project/ProjectConst.java
@@ -1,0 +1,4 @@
+package de.samply.project;
+
+public class ProjectConst {
+}