+++ src/main/java/de/samply/project/ProjectEvent.java
@@ -1,0 +1,12 @@
+package de.samply.project;
+
+public enum ProjectEvent {
+    CREATE,
+    ACCEPT,
+    REJECT,
+    ARCHIVE,
+    START_DEVELOP,
+    START_PILOT,
+    START_FINAL,
+    FINISH
+}