+++ src/main/java/de/samply/project/ProjectState.java
@@ -1,0 +1,13 @@
+package de.samply.project;
+
+public enum ProjectState {
+    DRAFT,
+    CREATED,
+    ACCEPTED,
+    DEVELOP,
+    PILOT,
+    FINAL,
+    FINISHED,
+    REJECTED,
+    ARCHIVED
+}