+++ CHANGELOG.md
@@ -8,3 +8,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - First version of the project
 - Spring Application
+- State Machine for project
+- Project states and event states